feat: add contributors script
