+++ pwa/package.json
@@ -20,6 +20,7 @@
     "@tailwindcss/typography": "^0.5.9",
     "classnames": "^2.3.2",
     "dayjs": "^1.11.8",
+    "dotenv": "^16.3.1",
     "gray-matter": "^4.0.3",
     "gsap": "^3.11.5",
     "lodash.throttle": "^4.1.1",
